refactor firestore util
