+++ src/app/api/user/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { AddUserRequest } from "@/lib/firebase/firestore/types";
+import { User, newUser, AddUserRequest } from "@/lib/firebase/firestore/types";
 import { addUser } from "@/lib/firebase/firestore/user/userUtil";
 
 /*
@@ -9,11 +9,12 @@ import { addUser } from "@/lib/firebase/firestore/user/userUtil";
  * Params:
  *  None
  * Request body:
+ *  user_id: global id of a User
  *  first: first name
  *  last: last name,
  *  email_address: email address
- *  user_id: global id of a User
- *  pfp: path to pfp image [optional]
+ *  phone_number?: phone number
+ *  pfp?: path to pfp image
  * Return:
  *  data: id of the new User generated by firestore
  *  error: error or null
@@ -28,7 +29,18 @@ export async function POST(req: Request) {
       throw new Error("missing required fields");
     }
 
-    const id: string = await addUser(data);
+    const user_id: string = data.user_id;
+    const user: User = newUser();
+    user.first = data.first;
+    user.last = data.last;
+    user.email_address = data.email_address;
+    if (data.phone_number) {
+      user.phone_number = data.phone_number;
+    }
+    if (data.pfp) {
+      user.pfp = data.pfp;
+    }
+    const id: string = await addUser(user_id, user);
 
     return NextResponse.json({ data: { user_id: id }, error: null });
   } catch (e: unknown) {
